working

+++ src/forestTypes/rerf/fpRerFBase.h
@@ -119,7 +119,7 @@ namespace fp {
 				}
 				std::cout << "\nnumWrong= " << numWrong << "\n";
 
-				return numWrong/numTried;
+				return (float)numWrong/(float)numTried;
 			}
 	};
